fix: improve table UX and harden error handling
