+++ src/app/staff/settings/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function SettingsPage(): never {
+  redirect("/staff/settings/skills");
+}
